requirements.txt
discord.py
yfinance
python-dotenv
pandas
robin_stocks
matplotlib
curl_cffi